requirements.txt
streamlit>=1.30
numpy>=1.22
pandas>=1.5
rich>=13.0
typer>=0.9
pypdf>=4.0
faiss-cpu>=1.8.0.post1
sentence-transformers>=2.2.2
scikit-learn>=1.3
joblib>=1.3
